+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # 0.1.1
 - typo fix: `MCaptcha::decrement_visiotr()` became `MCaptcha::decrement_visitor()`
 - `serde::{Serialize, Deserialize}` impls (shouldn't break anything)
+- `MCaptcha` throws error when duration is 0
+- `Visitor` is changed to `AddVisitor`